requirements.txt
requests
pandas
click
python-dotenv